+++ src/useForm.ts
@@ -2,17 +2,9 @@ import React from 'react';
 
 import { createFormControl } from './logic/createFormControl';
 import getProxyFormState from './logic/getProxyFormState';
-import shouldRenderFormState from './logic/shouldRenderFormState';
 import deepEqual from './utils/deepEqual';
 import isFunction from './utils/isFunction';
-import {
-  FieldValues,
-  FormState,
-  InternalFieldName,
-  UseFormProps,
-  UseFormReturn,
-} from './types';
-import { useSubscribe } from './useSubscribe';
+import { FieldValues, FormState, UseFormProps, UseFormReturn } from './types';
 
 /**
  * Custom hook to manage the entire form.
@@ -75,7 +67,7 @@ export function useForm<
 
   if (!_formControl.current) {
     _formControl.current = {
-      ...createFormControl(props),
+      ...(props.control ? props.control : createFormControl(props)),
       formState,
     };
   }
@@ -83,23 +75,15 @@ export function useForm<
   const control = _formControl.current.control;
   control._options = props;
 
-  useSubscribe({
-    subject: control._subjects.state,
-    next: (
-      value: Partial<FormState<TFieldValues>> & { name?: InternalFieldName },
-    ) => {
-      if (
-        shouldRenderFormState(
-          value,
-          control._proxyFormState,
-          control._updateFormState,
-          true,
-        )
-      ) {
-        updateFormState({ ...control._formState });
-      }
-    },
-  });
+  React.useEffect(
+    () =>
+      control._subscribe({
+        formState: control._proxyFormState,
+        callback: () => updateFormState({ ...control._formState }),
+        reRenderRoot: true,
+      }),
+    [control],
+  );
 
   React.useEffect(
     () => control._disableForm(props.disabled),
@@ -135,7 +119,7 @@ export function useForm<
 
   React.useEffect(() => {
     if (!control._state.mount) {
-      control._updateValid();
+      control._setValid();
       control._state.mount = true;
     }
 
@@ -149,7 +133,7 @@ export function useForm<
 
   React.useEffect(() => {
     props.shouldUnregister &&
-      control._subjects.values.next({
+      control._subjects.state.next({
         values: control._getWatch(),
       });
   }, [props.shouldUnregister, control]);