+++ src/useForm.ts
@@ -154,6 +154,12 @@ export function useForm<
       });
   }, [props.shouldUnregister, control]);
 
+  React.useEffect(() => {
+    if (_formControl.current) {
+      _formControl.current.watch = _formControl.current.watch.bind({});
+    }
+  }, [formState]);
+
   _formControl.current.formState = getProxyFormState(formState, control);
 
   return _formControl.current;