+++ CHANGELOG.md
@@ -17,7 +17,9 @@
 - added 'validateFields' to formState
 
 ```tsx
-const { formState: { validateFields } } = useForm();
+const {
+  formState: { validateFields },
+} = useForm();
 ```
 
 ## [7.49.0] - 2023-12-10
@@ -342,7 +344,8 @@ function App() {
   return (
     <form>
       <CustomFormComponent control={control} name="foo" /> {/* no error */}
-      <CustomFormComponent control={control} name="baz" /> {/*  throw an error since baz is string */}
+      <CustomFormComponent control={control} name="baz" />{' '}
+      {/*  throw an error since baz is string */}
     </form>
   );
 }
@@ -825,7 +828,8 @@ const App = () => {
           },
         })}
       />
-      <input {...register('lastName', { deps: ['firstName'] })} /> // dependant validation
+      <input {...register('lastName', { deps: ['firstName'] })} /> // dependant
+      validation
     </form>
   );
 };