+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [7.53.0] - 2024-8-31
+
+### Added
+
+- add support for `onBlur` with formState `isValid`
+
+### Changed
+
+- `validateFields` will only trigger re-render for async validation
+
 ## [7.51.0] - 2024-3-2
 
 ### Added
@@ -7,7 +17,9 @@
 - added 'validateFields' to formState
 
 ```tsx
-const { formState: { validateFields } } = useForm();
+const {
+  formState: { validateFields },
+} = useForm();
 ```
 
 ## [7.49.0] - 2023-12-10
@@ -332,7 +344,8 @@ function App() {
   return (
     <form>
       <CustomFormComponent control={control} name="foo" /> {/* no error */}
-      <CustomFormComponent control={control} name="baz" /> {/*  throw an error since baz is string */}
+      <CustomFormComponent control={control} name="baz" />{' '}
+      {/*  throw an error since baz is string */}
     </form>
   );
 }
@@ -815,7 +828,8 @@ const App = () => {
           },
         })}
       />
-      <input {...register('lastName', { deps: ['firstName'] })} /> // dependant validation
+      <input {...register('lastName', { deps: ['firstName'] })} /> // dependant
+      validation
     </form>
   );
 };